+++ src/i18n/locale-ca.json
@@ -929,5 +929,6 @@
     "MEMBER_PAGE_PAYMENT_METHOD": "Pagament",
     "MEMBER_PAGE_PAYMENT_METHOD_QUESTION": "Escull el mètode de pagament",
     "MEMBER_PAGE_SUMMARY": "Revisa i confirma les dades",
-    "MEMBER_PAGE_SUMMARY_QUESTION": "Revisa les dades i, si són correctes, clica el botó \"Finalitzar\""
+    "MEMBER_PAGE_SUMMARY_QUESTION": "Revisa les dades i, si són correctes, clica el botó \"Finalitzar\"",
+    "CIF": "CIF"
 }
